test(utils): move objectToCliArgs to transform, remove fs from executeProcess
