+++ packages/utils/src/lib/execute-process.ts
@@ -168,66 +168,3 @@ export function executeProcess(cfg: ProcessConfig): Promise<ProcessResult> {
     });
   });
 }
-
-type ArgumentValue = number | string | boolean | string[];
-export type CliArgsObject<T extends object = Record<string, ArgumentValue>> =
-  T extends never
-    ? Record<string, ArgumentValue | undefined> | { _: string }
-    : T;
-
-/**
- * Converts an object with different types of values into an array of command-line arguments.
- *
- * @example
- * const args = objectToProcessArgs({
- *   _: ['node', 'index.js'], // node index.js
- *   name: 'Juanita', // --name=Juanita
- *   formats: ['json', 'md'] // --format=json --format=md
- * });
- */
-export function objectToCliArgs<
-  T extends object = Record<string, ArgumentValue>,
->(params?: CliArgsObject<T>): string[] {
-  if (!params) {
-    return [];
-  }
-
-  return Object.entries(params).flatMap(([key, value]) => {
-    // process/file/script
-    if (key === '_') {
-      if (Array.isArray(value)) {
-        return value;
-      } else {
-        return [value + ''];
-      }
-    }
-    const prefix = key.length === 1 ? '-' : '--';
-    // "-*" arguments (shorthands)
-    if (Array.isArray(value)) {
-      return value.map(v => `${prefix}${key}="${v}"`);
-    }
-    // "--*" arguments ==========
-
-    if (Array.isArray(value)) {
-      return value.map(v => `${prefix}${key}="${v}"`);
-    }
-
-    if (typeof value === 'string') {
-      return [`${prefix}${key}="${value}"`];
-    }
-
-    if (typeof value === 'number') {
-      return [`${prefix}${key}=${value}`];
-    }
-
-    if (typeof value === 'boolean') {
-      return [`${prefix}${value ? '' : 'no-'}${key}`];
-    }
-
-    // @TODO add support for nested objects `persist.filename`
-
-    throw new Error(`Unsupported type ${typeof value} for key ${key}`);
-  });
-}
-
-objectToCliArgs<{ z: number }>({ z: 5 });
