+++ packages/utils/src/lib/scoring.ts
@@ -7,7 +7,7 @@ import {
   PluginReport,
   Report,
 } from '@code-pushup/models';
-import { deepClone } from './transformation';
+import { deepClone } from './transform';
 
 export type EnrichedAuditReport = AuditReport & { plugin: string };
 export type WeighedAuditReport = EnrichedAuditReport & { weight: number };
