+++ packages/utils/src/lib/transform.unit.test.ts
@@ -4,10 +4,11 @@ import {
   deepClone,
   distinct,
   factorOf,
+  objectToCliArgs,
   objectToEntries,
   objectToKeys,
   toArray,
-} from './transformation';
+} from './transform';
 
 describe('toArray', () => {
   it('should transform non-array value into array with single value', () => {
@@ -86,13 +87,10 @@ describe('deepClone', () => {
       e: undefined,
     };
     const cloned = deepClone(obj);
-    expect(cloned).toEqual(obj);
+    expect(cloned).toStrictEqual(obj);
     expect(cloned).not.toBe(obj);
-    expect(cloned.c).toEqual(obj.c);
     expect(cloned.c).not.toBe(obj.c);
-    expect(cloned.c[0]).toEqual(obj.c[0]);
     expect(cloned.c[0]).not.toBe(obj.c[0]);
-    expect(cloned.c[1]).toEqual(obj.c[1]);
     expect(cloned.c[1]).not.toBe(obj.c[1]);
     expect(cloned.d).toBe(obj.d);
     expect(cloned.e).toBe(obj.e);
@@ -107,5 +105,62 @@ describe('factorOf', () => {
     [[1, 1], 1],
   ])('should return correct factor items', (items, factor) => {
     expect(factorOf(items, i => i < 1)).toEqual(factor);
+  });
+});
+
+describe('objectToCliArgs', () => {
+  it('should handle the "_" argument as script', () => {
+    const params = { _: 'bin.js' };
+    const result = objectToCliArgs(params);
+    expect(result).toEqual(['bin.js']);
+  });
+
+  it('should handle the "_" argument with multiple values', () => {
+    const params = { _: ['bin.js', '--help'] };
+    const result = objectToCliArgs(params);
+    expect(result).toEqual(['bin.js', '--help']);
+  });
+
+  it('should handle shorthands arguments', () => {
+    const params = {
+      e: `test`,
+    };
+    const result = objectToCliArgs(params);
+    expect(result).toEqual([`-e="${params.e}"`]);
+  });
+
+  it('should handle string arguments', () => {
+    const params = { name: 'Juanita' };
+    const result = objectToCliArgs(params);
+    expect(result).toEqual(['--name="Juanita"']);
+  });
+
+  it('should handle number arguments', () => {
+    const params = { parallel: 5 };
+    const result = objectToCliArgs(params);
+    expect(result).toEqual(['--parallel=5']);
+  });
+
+  it('should handle boolean arguments', () => {
+    const params = { progress: true };
+    const result = objectToCliArgs(params);
+    expect(result).toEqual(['--progress']);
+  });
+
+  it('should handle negated boolean arguments', () => {
+    const params = { progress: false };
+    const result = objectToCliArgs(params);
+    expect(result).toEqual(['--no-progress']);
+  });
+
+  it('should handle array of string arguments', () => {
+    const params = { format: ['json', 'md'] };
+    const result = objectToCliArgs(params);
+    expect(result).toEqual(['--format="json"', '--format="md"']);
+  });
+
+  it('should throw error for unsupported type', () => {
+    const params = { unsupported: undefined as any };
+    expect(() => objectToCliArgs(params)).toThrow('Unsupported type');
   });
 });
