+++ packages/utils/test/process-helper.mock.ts
@@ -14,18 +14,15 @@ export function getAsyncProcessRunnerConfig(
     throwError?: boolean;
     interval?: number;
     runs?: number;
-    outputFile?: string;
   } = { throwError: false },
 ) {
-  const outputFile = cfg?.outputFile || './tmp/out-async-runner.json';
   const args = [
     asyncProcessPath,
     cfg?.interval ? cfg.interval + '' : '10',
     cfg?.runs ? cfg.runs + '' : '4',
     cfg?.throwError ? '1' : '0',
-    outputFile,
   ];
-  return { command: 'node', args, outputFile };
+  return { command: 'node', args };
 }
 
 export function mockProcessConfig(
@@ -35,7 +32,6 @@ export function mockProcessConfig(
     ...{
       command: 'dummy-string',
       args: [],
-      outputFile: join('tmp', 'out.json'),
     },
     ...processConfig,
     observer: spyObserver(),
